refactor to move subscription logic to different service class
